requirements.txt
fastapi
langchain
python-dotenv
pydantic
langchain-openai
langchain-community
faiss-cpu